backend/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
distro==1.4.0
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
entrypoints==0.3
filelock==3.0.12
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
isort==5.5.1
keyring==18.0.1
keyrings.alt==3.4.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
more-itertools==4.2.0
numpy==1.17.4
oauthlib==3.1.0
packaging==20.3
pexpect==4.6.0
Pillow==7.0.0
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pytz==2020.1
pyxdg==0.26
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.8
virtualenv==20.0.17
